fix(test_host): make the host suite green, plain and under ASan/UBSan
- iambic: release debounce blanked the paddles at t=0 because the
  release timestamps started at 0; park them one blanking period in
  the past on init/reset. Also compute the Mode B release progress in
  int64 so a late release cannot wrap below the memory window start.
- cwnet_frame: explicit uint16_t cast on payload_len (-Wconversion
  under UBSan instrumentation).
- test_stream: the lag test used stream_push(), which compresses
  identical samples into idle ticks; use stream_push_raw().
- test_fault: fault_clear() keeps the lifetime counter as documented
  in fault.h; the test expected a reset.

+++ components/keyer_cwnet/src/cwnet_frame.c
@@ -121,7 +121,7 @@ cwnet_parse_result_t cwnet_frame_parse(cwnet_frame_parser_t *parser,
             case CWNET_PARSER_STATE_LENGTH_2: {
                 /* Long block - second length byte (little-endian) */
                 uint16_t len_high = (uint16_t)data[pos];
-                parser->payload_len = parser->length_byte_1 | (len_high << 8);
+                parser->payload_len = (uint16_t)(parser->length_byte_1 | (len_high << 8));
                 pos++;
                 parser->payload_received = 0;
